Couple fixes relating to the sphere test scene
First, we were incorrectly rotating around the object space normal instead of
the world space normal. This happened to work in scenes with uniform scaling,
but broke down when non-uniform scales were involved. Fixed this by converting
the normals into world space before computing the new player rotation. This
fixed an issue where the player would leave the plane of the sphere, causing
the "between points" determination to eventually fail.

Also tried reprojecting the position and direction onto the current face on
every iteration, just to snuff out any further problems with floating point
precision. This will have a side effect of masking most issues relating to
incorrect move directions, so hopefully there aren't any more major broken
cases remaining :-)

Finally, identified a degenerate case when the point is on an edge and the
movement direction is perfectly along that edge. This should be exceedingly
rare in normal gameplay (would have to look directly along the edge), and
should be exceedingly difficult not to work around (i.e. the player would have
to not change their look direction at all), so I'm going to ignore this case
for now. Worked around this in the test scene by slightly rotating the sphere.